Remove duplicated code formatting multi errors
