+++ cli/cmd/start.go
@@ -20,14 +20,13 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/docker/compose-cli/errdefs"
-
+	"github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
-	"github.com/hashicorp/go-multierror"
-
 	"github.com/docker/compose-cli/api/client"
+	"github.com/docker/compose-cli/cli/formatter"
+	"github.com/docker/compose-cli/errdefs"
 )
 
 // StartCommand starts containers
@@ -63,9 +62,6 @@ func runStart(ctx context.Context, args []string) error {
 		}
 		fmt.Println(id)
 	}
-	if errs != nil {
-		errs.ErrorFormat = formatErrors
-	}
-
+	formatter.SetMultiErrorFormat(errs)
 	return errs.ErrorOrNil()
 }
